chore: move interfaces
