+++ pyproject.toml
@@ -30,6 +30,12 @@ types-aioboto3 = {extras = ["dynamodb"], version = "^13.4.0"}
 mypy = "^1.15.0"
 ruff = "^0.9.6"
 
+[tool.ruff]
+include = ["pyproject.toml", "pynamodb/**/*.py"]
+
+[tool.ruff.lint.per-file-ignores]
+"*.py" = ["E721"]
+
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
